Foi implementado a regra de negócio onde empresa pode ver tudo de todos os setores e gestores podem gerenciar coisas apenas dos seus setores e funcionários visualizarem apenas itens do seu setor. Está regra de negócio só não é aplicad aa competências já que ela se refere a todas áreas da empresa.
Alterado posição do logo na tela inicial.

+++ Insignia.Painel/Controllers/SetoresController.cs
@@ -113,13 +113,20 @@ namespace Insignia.Painel.Controllers
 
             if (SetorModel != null)
             {
-                if (SetoresDAO.Remover(SetorModel.ID))
+                if (!SetoresDAO.PodeRemover(SetorModel.ID))
                 {
-                    return RedirectToAction("Listar");
+                    if (SetoresDAO.Remover(SetorModel.ID))
+                    {
+                        return RedirectToAction("Listar");
+                    }
+                    else
+                    {
+                        ViewBag.Error = "Ocorreu um erro ao tentar excluir o resgistro, favor entrar em contato com o administrador do sistema";
+                    }
                 }
                 else
                 {
-                    ViewBag.Error = "Ocorreu um erro ao tentar excluir o resgistro, favor entrar em contato com o administrador do sistema";
+                    ViewBag.Error = "Não é possível remover o setor " + SetorModel.Nome + ", pois existem badges cadastras para o mesmo.";
                 }
             }
